push ms bfs 8 lane branch

+++ src/processor/operator/recursive_extend/recursive_join.cpp
@@ -332,7 +332,7 @@ void RecursiveJoin::computeBFSOneThreadOneMorsel(ExecutionContext* context) {
 
 void RecursiveJoin::computeMSBFSMorsel(kuzu::processor::ExecutionContext* context) {
     auto msBFSMorsel = (reinterpret_cast<MSBFSMorsel<false>*>(bfsMorsel.get()));
-    uint64_t *temp, *x = msBFSMorsel->visit, *next_ = msBFSMorsel->next;
+    uint8_t *temp, *x = msBFSMorsel->visit, *next_ = msBFSMorsel->next;
     while (doMSBFS(msBFSMorsel->seen, x, next_, msBFSMorsel->maxOffset, context)) {
         msBFSMorsel->updateBFSLevel();
         temp = x;
@@ -341,7 +341,7 @@ void RecursiveJoin::computeMSBFSMorsel(kuzu::processor::ExecutionContext* contex
     }
 }
 
-bool RecursiveJoin::doMSBFS(uint64_t* seen, uint64_t* curFrontier, uint64_t* nextFrontier,
+bool RecursiveJoin::doMSBFS(uint8_t* seen, uint8_t* curFrontier, uint8_t* nextFrontier,
     uint64_t maxOffset, kuzu::processor::ExecutionContext* context) {
     for (auto offset = 0u; offset < (maxOffset + 1); offset++) {
         seen[offset] |= curFrontier[offset];
@@ -361,8 +361,8 @@ bool RecursiveJoin::doMSBFS(uint64_t* seen, uint64_t* curFrontier, uint64_t* nex
     return active;
 }
 
-void RecursiveJoin::callMSBFSRecursivePlan(const uint64_t* seen, const uint64_t* curFrontier,
-    uint64_t* nextFrontier, common::offset_t parentOffset, bool& isBFSActive,
+void RecursiveJoin::callMSBFSRecursivePlan(const uint8_t* seen, const uint8_t* curFrontier,
+    uint8_t* nextFrontier, common::offset_t parentOffset, bool& isBFSActive,
     kuzu::processor::ExecutionContext* context) {
     scanFrontier->setNodeID(common::nodeID_t{parentOffset, *begin(dataInfo->dstNodeTableIDs)});
     while (recursiveRoot->getNextTuple(context)) {
